Extract printClass and printTypeParameters (#9713)
